Migration vers Composer
* PSR-0
* Package Composer
* Ajout du certificat SSL de Terena
* Mise à jour de l'exemple (dans le README)
