+++ Payutc/Client/JsonClient.php
@@ -1,36 +1,27 @@
-<?php namespace JsonClient;
+<?php
+/**
+* payutc
+* Copyright (C) 2013-2014 payutc <payutc@assos.utc.fr>
+*
+* This file is part of payutc
+* 
+* payutc is free software: you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation, either version 3 of the License, or
+* (at your option) any later version.
+*
+* payutc is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License
+* along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
 
+namespace Payutc\Client;
 
-use \JsonClient\JsonException;
-use \JsonClient\JsonClient;
-
-class JsonException extends \Exception
-{
-	protected $type;
-	protected $http_code;
-	
-	public function __construct($type, $code, $msg = null, $http_code = null)
-	{
-		parent::__construct($msg, $code);
-		$this->type = $type;
-		$this->http_code = $http_code;
-	}
-	
-	public function getType()
-	{
-		return $this->type;
-	}
-	
-	public function getHttpCode()
-	{
-		return $this->http_code;
-	}
-	
-	public function __tostring()
-	{
-		return "JsonException(type={$this->type}, code={$this->getCode()}, msg={$this->getMessage()}, http_code={$this->http_code})";
-	}
-}
+use \Payutc\Client\JsonException;
 
 /**
  * Basic client to do POST and GET
@@ -54,6 +45,7 @@ class JsonClient
 				CURLOPT_USERAGENT => $this->useragent,
 				CURLOPT_RETURNTRANSFER => true,
 				CURLOPT_SSL_VERIFYPEER => true,
+                CURLOPT_CAINFO => __DIR__ . "/../../TERENA_SSL_CA.pem",
                 CURLOPT_HEADERFUNCTION => array($this, 'readHeader'),
 			);
 	}
@@ -129,29 +121,3 @@ class JsonClient
 		}
 	}
 }
-
-
-/**
- * Auto generate calls client:
- * 
- * $client->myFunc(array("a"=>42)) will remotly call method "myFunc" with 
- * param a=42.
- * 
- */
-class AutoJsonClient extends JsonClient
-{
-	public function __call($name, $arguments)
-	{
-		if (count($arguments) < 1) {
-			return $this->apiCall($name);
-		}
-		else {
-			return $this->apiCall($name, $arguments[0]);
-		}
-	}
-}
-
-
-
-
-
