+++ composer.json
@@ -1,0 +1,24 @@
+{
+    "name": "payutc/php-client",
+    "type": "library",
+    "description": "PHP client for the payutc servers",
+    "keywords": ["simde","utc","payutc"],
+    "homepage": "https://github.com/payutc/php-client",
+    "license": "GPLv3",
+    "authors": [
+        {
+            "name": "payutc",
+            "email": "payutc@assos.utc.fr",
+            "homepage": "http://payutc.github.io",
+            "role": "Owner"
+        }
+    ],
+    "require": {
+        "php": ">=5.3.0"
+    },
+    "autoload": {
+        "psr-0": {
+            "Payutc": "."
+        }
+    }
+}
